Pull out KVNamespace, ProjectType Project submodules
