+++ src/settings/project/mod.rs
@@ -1,10 +1,14 @@
+mod kv_namespace;
+mod project_type;
+
+pub use kv_namespace::KVNamespace;
+pub use project_type::ProjectType;
+
 use crate::terminal::emoji;
 
 use std::collections::HashMap;
-use std::fmt;
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::str::FromStr;
 
 use log::info;
 
@@ -24,62 +28,6 @@ pub struct Project {
     pub routes: Option<HashMap<String, String>>,
     #[serde(rename = "kv-namespaces")]
     pub kv_namespaces: Option<Vec<KVNamespace>>,
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-#[serde(rename_all = "lowercase")]
-pub enum ProjectType {
-    JavaScript,
-    Rust,
-    Webpack,
-}
-
-impl Default for ProjectType {
-    fn default() -> Self {
-        ProjectType::Webpack
-    }
-}
-
-impl fmt::Display for ProjectType {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let printable = match *self {
-            ProjectType::JavaScript => "js",
-            ProjectType::Rust => "rust",
-            ProjectType::Webpack => "webpack",
-        };
-        write!(f, "{}", printable)
-    }
-}
-
-impl FromStr for ProjectType {
-    type Err = failure::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        match s {
-            "javascript" => Ok(ProjectType::JavaScript),
-            "rust" => Ok(ProjectType::Rust),
-            "webpack" => Ok(ProjectType::Webpack),
-            _ => failure::bail!("{} is not a valid wrangler project type!", s),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Deserialize, Serialize)]
-pub struct KVNamespace {
-    id: String,
-    binding: String,
-}
-
-impl fmt::Display for KVNamespace {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "id: {}, binding: {}", self.id, self.binding)
-    }
-}
-
-impl std::cmp::PartialEq for KVNamespace {
-    fn eq(&self, other: &Self) -> bool {
-        self.id == other.id && self.binding == other.binding
-    }
 }
 
 impl Project {
